[Chore] rename bunch of files by removing postfix `Extension`
