+++ assets/locale/zh-cn.json
@@ -19,9 +19,6 @@
             "rendering": "渲染中...",
             "error": "错误公式"
         },
-        "editor/contrib/blockPlaceHolderExtension/blockPlaceHolderExtension": {
-            "emptyBlockPlaceHolder": "开始输入，或者试试 '@' 触发 AI，'/' 呼出命令..."
-        },
         "editor/common/markdown": {
             "code": "代码",
             "heading": "标题",
@@ -36,6 +33,9 @@
         },
         "editor/model/documentNode/node/html/html": {
             "empty": "空HTML"
+        },
+        "editor/contrib/blockPlaceHolder/blockPlaceHolder": {
+            "emptyBlockPlaceHolder": "开始输入，或者试试 '@' 触发 AI，'/' 呼出命令..."
         }
     }
 }
