+++ src/editor/contrib/blockHandle/blockHandle.scss
