+++ src/editor/contrib/blockPlaceHolder/blockPlaceHolder.scss
