+++ src/editor/contrib/blockPlaceHolder/blockPlaceHolder.ts
@@ -1,9 +1,9 @@
-import "src/editor/contrib/blockPlaceHolderExtension/blockPlaceHolder.scss";
+import "src/editor/contrib/blockPlaceHolder/blockPlaceHolder.scss";
 import { EditorExtension, IEditorExtension } from "src/editor/common/editorExtension";
 import { ProseTools } from "src/editor/common/proseUtility";
 import { EditorExtensionIDs } from "src/editor/contrib/builtInExtensionList";
 import { IEditorWidget } from "src/editor/editorWidget";
-import { ProseDecoration, ProseDecorationSet, ProseDecorationSource, ProseEditorState, ProseEditorView } from "src/editor/common/proseMirror";
+import { ProseDecoration, ProseDecorationSet, ProseDecorationSource, ProseEditorState } from "src/editor/common/proseMirror";
 import { I18nService, II18nService } from "src/platform/i18n/browser/i18nService";
 
 interface IEditorBlockPlaceHolderExtension extends IEditorExtension {
