+++ src/editor/contrib/dragAndDrop/dropBlinkRenderer.scss
