+++ src/editor/contrib/dragAndDrop/scrollOnEdgeController.ts
