+++ src/editor/contrib/slashCommand/slashCommand.scss
